[ADD] RenewToken

+++ app/Http/Controllers/Products.php
@@ -14,29 +14,29 @@ class Products extends Controller
     public function index(Request $request)
     {
         if ($request->user()->role_id !== 1) {
-            $products = Product::where('user_id', $request->user()->id)->get();
+
+            $products = Product::with('user')
+                ->where('user_id', $request->user()->id)
+                ->get();
         } else {
-            $products = Product::all();
+            $products = Product::with('user')
+                ->get();
         }
 
         foreach ($products as $product) {
 
             if (!$product->image) {
-                $product->image_base64 = null;
+                $product->imageUrl = null;
                 continue;
             }
         
-            $imagePath = storage_path('app/public/' . $product->image);
-
-            if (file_exists($imagePath)) {
-                $imageContent = file_get_contents($imagePath);
-                $product->image_base64 = base64_encode($imageContent);
-            } else {
-                $product->image_base64 = null;
-            }
+            $product->image_url = url( 'storage/'.$product->image);
         }
 
-        return response()->json($products);
+        return response()->json([
+            'message' => 'Productos Obtenidos',
+            'data' => $products,
+        ], 200);
     }
 
     /**
@@ -44,6 +44,11 @@ class Products extends Controller
      */
     public function store(Request $request)
     {
+
+        $request->validate([
+            'name' => 'required',
+            'price' => 'required',
+        ]);
 
         if ($request->hasFile('image')) {
             $imagePath = $request->file('image')->store('products', 'public');
@@ -58,9 +63,9 @@ class Products extends Controller
         $product->save();
 
         return response()->json([
-            'message' => 'Product created successfully',
+            'message' => 'Product Creado',
             'data' => null,
-        ]);
+        ], 200);
     }
 
     /**
@@ -76,39 +81,50 @@ class Products extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $request->validate([
+            'name' => 'required',
+            'price' => 'required',
+        ]);
 
-        if($request->user()->role_id !== 1) {
+        if ($request->user()->role_id !== 1) {
             $product = Product::where('user_id', $request->user()->id)->find($id);
 
             if (!$product) {
                 return response()->json([
-                    'message' => 'Product not found',
+                    'message' => 'Producto no encontrado',
                     'data' => null,
-                ]);
+                ], 400);
             }
-
         } else {
             $product = Product::find($id);
+
+            if (!$product) {
+                return response()->json([
+                    'message' => 'Producto no encontrado',
+                    'data' => null,
+                ], 400);
+            }
         }
 
-        if ($request->hasFile('image')) {
+        if ($request->hasFile('image') && $request->file('image')->isValid()) {
             if ($product->image) {
                 Storage::disk('public')->delete($product->image);
             }
             $imagePath = $request->file('image')->store('products', 'public');
         }
-    
+
         $product->name = $request->name;
         $product->description = $request->description;
         $product->price = $request->price;
         $product->image = $imagePath ?? $product->image;
         $product->save();
-        
+
         return response()->json([
-            'message' => 'Product updated successfully',
+            'message' => 'Producto actualizado',
             'data' => null,
-        ]);
+        ], 200);
     }
+
 
     /**
      * Remove the specified resource from storage.
@@ -116,10 +132,20 @@ class Products extends Controller
     public function destroy(string $id)
     {
         $product = Product::find($id);
+
+        if (!$product) {
+            return response()->json([
+                'message' => 'Producto no encontrado',
+                'data' => null,
+            ], 400);
+        }
+
         $product->active = !$product->active;
+        $product->save();
+
         return response()->json([
-            'message' => 'Product deleted successfully',
+            'message' => 'Estado del producto actualizado',
             'data' => null,
-        ]);
+        ], 200);
     }
 }
